requirements.txt
fastapi
uvicorn
python-dotenv
pymupdf
pinecone
easyocr
python-pptx
docx2txt
Pillow
numpy
sentence-transformers
langchain
langchain-core
langchain-community
langchain-text-splitters
python-magic
python-multipart
requests
rank-bm25